Change diff highlighting to better fit the theme
Highlighting diffs with bright background colors may look good on bright
websites like github, but it looks really bad on dark themed websites,
especially neonish-with-dark-background websites like this one.

Change the diff highlighting to pure green/red on default background for
added/removed lines, and bold for the specific differing part of the lines.

At the moment none of the 2 highlighters we're using supports the
specific-part-of-line thing, but maybe in the fututre...